requirements.txt

Flask==2.0.1
Flask-Cors==3.0.10
pycryptodome==3.11.0
psycopg2-binary==2.9.1
Werkzeug==2.0.2
